# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel-training docs",
        "description": "Api description...",
        "termsOfService": "",
        "contact": {
            "email": "contact@mysite.com"
        },
        "license": {
            "name": "Private License",
            "url": "URL to the license"
        },
        "version": "1.0.0"
    },
    "host": "laravel.local",
    "basePath": "/",
    "schemes": [
        "http",
        "https"
    ],
    "paths": {
        "/api/posts": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Get list of posts",
                "description": "Returns list of posts",
                "operationId": "getpostsList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/post/create": {
            "post": {
                "tags": [
                    "posts"
                ],
                "summary": "Create post",
                "description": "Create post",
                "operationId": "getpostsList",
                "responses": {
                    "201": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "posts"
                ],
                "summary": "Show post in detail",
                "description": "Returns detail of post",
                "operationId": "getpostsList",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "unauthenticated"
                    }
                },
                "security": [
                    {
                        "api_key_security_example": []
                    }
                ]
            }
        }
    },
    "definitions": {},
    "securityDefinitions": {
        "passport": {
            "type": "oauth2",
            "flow": "password",
            "tokenUrl": "/oauth/token",
            "scopes": {}
        }
    },
    "externalDocs": {
        "description": "Find out more about my website",
        "url": "http..."
    }
}